restructure filemanager folder
