+++ react-components/oodt_opsui_sample_app/src/components/fileManager/Product/Product.js
@@ -22,7 +22,7 @@ import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
 import { withStyles } from "@material-ui/core";
 import PropTypes from "prop-types";
-import { fmconnection } from "../../constants/fmconnection";
+import { fmconnection } from "constants/connection";
 import Grid from "@material-ui/core/Grid";
 import Fab from "@material-ui/core/Fab";
