+++ react-components/oodt_opsui_sample_app/src/components/fileManager/Product/index.js
